[Refactor] (UWB): 重构UWB任务和日志配置
将UWB任务实现移至CX310驱动文件
简化日志UART配置为uart7
减少LED任务栈大小并移除调试代码

优化代码结构，提高模块化程度
